requirements.txt
fastapi
uvicorn[standard]
pandas
sqlalchemy
python-dotenv
jinja2
pytest
python-multipart
requests